* commit '6fcb56f14513b9da18976b82bca7e2f8d486658b':
  PT-10214 - Raise controller test coverage

+++ Migration/Run/RunService.php
@@ -522,6 +522,10 @@ class RunService implements RunServiceInterface
             $existsIsoCodes[$currency->getIsoCode()] = $currency->getId();
         }
 
+        if (empty($currenciesToDelete)) {
+            return;
+        }
+
         $this->currencyRepository->delete($currenciesToDelete, $context);
     }
 }
